requirements.txt
streamlit
ultralytics
Pillow
pandas